+++ CloudShell.Abstractions/ResourceManager/ResourceTabGroupTitles.cs
@@ -4,6 +4,7 @@ public static class ResourceTabGroupTitles
 {
     public const string General = "General";
     public const string Application = "Application";
+    public const string Messaging = "Messaging";
     public const string Networking = "Networking";
     public const string Runtime = "Runtime";
     public const string Storage = "Storage";